Make MVR scene node reader self-contained
