+++ mvr/CMakeLists.txt
@@ -4,6 +4,9 @@ target_sources(${PROJECT_NAME} PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/mvrimporter.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/mvr_import_types.h
     ${CMAKE_CURRENT_SOURCE_DIR}/mvr_scene_node_reader.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/mvr_scene_node_reader.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/mvr_scene_node_reader_detail.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/mvr_scene_node_reader_support.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/mvr_import_package.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/mvr_merge_analyzer.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/mvr_merge_applier.cpp
