+++ tests/mvr_scene_node_reader_header_compile_test.cpp
@@ -1,0 +1,4 @@
+#include "mvr_scene_node_reader.h"
+
+// Confirms that the scene-node reader interface is independently includable.
+int main() { return 0; }
